% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun_R.R
\name{data}
\alias{data}
\title{failure machines data}
\format{
A data frame with 297 rows and 5 variables:
\describe{
  \item{D}{Temperature difference}
  \item{V}{speed}
  \item{N}{torque}
  \item{T}{service time}
  \item{L}{machine level}
}
}
\description{
A prediction variable data of failure machines
}
