% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun_R.R
\name{flcrm}
\alias{flcrm}
\title{Estimation of parameters for Functional Cox Model}
\usage{
flcrm(time, event, X, Z, hh = 40, k = 5)
}
\arguments{
\item{time}{survival time (vector)}

\item{event}{equals 1 if the observed event is a failure time and 0 otherwise (vector)}

\item{X}{functional factor (high-dimensional matrix)}

\item{Z}{scalar factor (matrix)}

\item{hh}{optional parameters to control the number of principal components (default 40)}

\item{k}{optional parameters to control the number of principal components (default 5)}
}
\value{
a list of estimation of parameters for \code{X} and \code{Z}
}
\description{
Using functional principal component analysis method to estimate the parameters of Functional Cox Model
}
\examples{
\dontrun{
data(time)
data(event)
data(X)
data(Z)
attach(time)
attach(event)
attach(X)
attach(Z)
res <- flcrm(time,event,X,Z)
}
}
