% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{Gibbs sampler in homework}
\usage{
gibbsC(mu1, sigma1, mu2, sigma2, rho, N)
}
\arguments{
\item{mu1}{mean}

\item{sigma1}{standard deviation}

\item{mu2}{mean}

\item{sigma2}{standard deviation}

\item{rho}{coefficient of correlation}

\item{N}{numbers}
}
\value{
a random sample of size \code{N}
}
\description{
Gibbs sampler in homework for two dimensional normal state
}
\examples{
\dontrun{
mu1=mu2=0
sigma1=sigma2=1
rho=0.9
N=100
gib<-gibbsC(mu1,sigma1,mu2,sigma2,rho,N)
}
}
