% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun_R.R
\name{mlearning}
\alias{mlearning}
\title{Integration of machine learning algorithms}
\usage{
mlearning(label, data, k = 5, methods = c("nnet", "knn", "SVM", "rF"))
}
\arguments{
\item{label}{response variable (logical)}

\item{data}{predictive Variables (numeric)}

\item{k}{k fold cross validation}

\item{methods}{choose machine learning algorithms}
}
\value{
a list given scores of used algorithms
}
\description{
Using k-fold cross validation combined with grid search to give various scores of nnet svm knn randomForest
}
\examples{
\dontrun{
data(label)
data(data)
attach(label)
attach(data)
res <- mlearning(label,data,methods=c("knn","rF"))
}
}
