% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleC}
\alias{scaleC}
\title{Scaling and Centering of Matrix-like Objects}
\usage{
scaleC(X, center, scale)
}
\arguments{
\item{X}{a numeric matrix}

\item{center}{centralization or not}

\item{scale}{standardization or not}
}
\value{
centered and scaled matrix
}
\description{
center and scale the columns of a numeric matrix.
}
\examples{
\dontrun{
X<-matrix(1:10,ncol=2)
res <- scaleC(X,center=T,scale=T)
}
}
