\name{StatComp22088-package}
\alias{StatComp22088-package}
\docType{package}
\title{
  This package provides several functions for Functional Cox Model and some machine learning methods.
}
\description{
  \code{flcrm} estimate the parameters of Functional Cox Model. \code{mlearning} use k-fold cross validation combined with grid search to give various scores of nnet, svm, knn and randomForest. \code{SVD_V} and \code{SVDS_V} calculate SVD decomposition using RcppArmadillo. \code{scaleC} scale and center a Matrix-like Objects using RcppArmadillo.
}
\author{
22088

Maintainer: 22088 <one@mail.ustc.edu.cn>
}
\keyword{ package }
