% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SVD_V}
\alias{SVD_V}
\title{SVD using RcppArmadillo}
\usage{
SVD_V(X)
}
\arguments{
\item{X}{a numeric matrix}
}
\value{
SVD Decomposition Right Matrix
}
\description{
A faster SVD than R
}
\examples{
\dontrun{
X<-matrix(1:10,ncol=2)
res <- SVD_V(X)
}
}
