% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SVDS_V}
\alias{SVDS_V}
\title{SVD using RcppArmadillo also}
\usage{
SVDS_V(X, k = 10L)
}
\arguments{
\item{X}{a sparse matrix}

\item{k}{number of expected singular values}
}
\value{
First k vectors of SVD Decomposition Right Matrix
}
\description{
First k singular values and vectors for sparse matrix
}
\examples{
\dontrun{
X<-matrix(sample(0:2,size=20,replace=T,prob = c(0.8,0.1,0.1)),ncol=4)
res <- SVDS_V(X)
}
}
