% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun_R.R
\name{X}
\alias{X}
\title{Hippocampal surface data}
\format{
A data frame with 300 rows and 373 individuals
}
\description{
Hippocampal surface data in ADNI
}
\details{
Complete data with 30000 rows are available in https://adni.loni.usc.edu/study-design/
}
